Small bug fixes + Refactoring (Didnt like the previous package names)
