+++ src/main/java/com/instanote/esp/EspApplication.java
@@ -1,13 +1,13 @@
-package com.ESP.Project;
+package com.instanote.esp;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class ProjectApplication {
+public class EspApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(ProjectApplication.class, args);
+		SpringApplication.run(EspApplication.class, args);
 	}
 
 }
