+++ src/main/java/com/instanote/esp/HomeController.java
@@ -1,4 +1,4 @@
-package com.ESP.Project;
+package com.instanote.esp;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
