+++ src/main/resources/templates/index.html
@@ -10,7 +10,7 @@
         <div class="container">
           <div class="upper-portion">
           <div class="navbar">
-            <div class="name">InstagramNote AI</div>
+            <div class="name">InstaNote AI</div>
             <div class="user">Hi, Username</div>
           </div>
             <div class="bar">
